backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
exceptiongroup==1.1.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==22.0
Pillow==9.4.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pyrsistent==0.19.3
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
simplejson==3.18.1
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.13
